requirements.txt
bintrees==2.0.7
click==6.7
gdax==1.0.6
gevent==1.2.2
greenlet==0.4.12
-e git+https://github.com/veox/python2-krakenex.git@b9407a24cb14756a7965c5b9c288b00db7711cca#egg=krakenex
marshmallow==2.13.5
pymongo==3.4.0
python-dateutil==2.6.1
requests==2.13.0
six==1.10.0
websocket-client==0.40.0
